cleaned up cli calling
