+++ .gitignore
@@ -122,3 +122,7 @@ dmypy.json
 
 # Pyre type checker
 .pyre/
+
+# extra
+.idea
+scaden.egg-info
