--- .idea/libraries/R_User_Library.xml
@@ -1,6 +1,0 @@
-<component name="libraryTable">
-  <library name="R User Library">
-    <CLASSES />
-    <SOURCES />
-  </library>
-</component>
