--- conda.recipe/meta.yaml
@@ -1,55 +1,0 @@
-{% set data = load_setup_py_data() %}
-
-package:
-  name: 'scaden'
-  version: {{ data.get('version') }}
-
-source:
-  path: ..
-
-build:
-  number: 0
-  script: pip install . --no-deps --ignore-installed --no-cache-dir
-  noarch: python
-  script_env:
-    - COVERAGE_DIR
-
-requirements:
-  host:
-    - python {{ python }}
-    - pip
-    - twine
-  run:
-    - python {{ python }}
-    - pandas
-    - numpy
-    - scikit-learn
-    - anndata
-    - tensorflow >=2.0
-    - scanpy
-    - tqdm
-    - click
-
-test:
-  requires:
-    - coverage
-#  source_files:
-#    - .coveragerc
-  commands:
-    - scaden --help
-
-
-# tests fail because of tesnorflow, commenting out
-#test:
-#  commands:
-#    - scaden --help
-
-about:
-  home: https://github.com/KevinMenden/scaden
-  license: MIT
-  summary: Cell type deconvolution using single cell data
-
-extra:
-  skip-lints:
-    - missing_tests
-
