+++ scaden/__init__.py
@@ -1,0 +1,3 @@
+import pkg_resources
+
+__version__ = pkg_resources.get_distribution("scaden").version
