--- scripts/scaden
@@ -1,14 +1,0 @@
-#!/usr/bin/env python
-
-"""
-cdn Command Line Interface
-author: Kevin Menden, DZNE Tübingen
-
-This is the main file for executing the cdn program.
-"""
-
-from scaden.cli import main
-
-if __name__ == '__main__':
-    main()
-
